+++ .github/workflows/cypress-testing.yml
@@ -96,7 +96,7 @@ jobs:
           git clone https://github.com/glific/cypress-testing.git
           echo done. go to dir.
           cd cypress-testing
-          git checkout registration_form
+          git checkout main
           cd ..
           cp -r cypress-testing/cypress cypress
           yarn add cypress@13.6.2
