+++ src/containers/Organization/RegistrationForm/FormLayout/FormLayout.tsx
@@ -176,8 +176,6 @@ export const FormLayout = ({
             variant="contained"
             color="primary"
             onClick={() => {
-              console.log(formik.errors);
-
               onSaveButtonClick(formik.errors);
               formik.submitForm();
             }}
